execinfrapb: break the dependency on sem/builtins
This commit moves several utility functions from `execinfrapb` into
`execinfra` package in order to break the dependency of the former on
`sem/builtins` (which eventually depends on the `c-deps`).

Release note: None
