+++ pkg/sql/colexec/colexecagg/aggregators_util.go
@@ -13,6 +13,7 @@ package colexecagg
 import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colexecop"
 	"github.com/cockroachdb/cockroach/pkg/sql/colmem"
+	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
@@ -30,7 +31,7 @@ type NewAggregatorArgs struct {
 	InputTypes     []*types.T
 	Spec           *execinfrapb.AggregatorSpec
 	EvalCtx        *tree.EvalContext
-	Constructors   []execinfrapb.AggregateConstructor
+	Constructors   []execinfra.AggregateConstructor
 	ConstArguments []tree.Datums
 	OutputTypes    []*types.T
 }
