+++ pkg/sql/colexec/colexecagg/ordered_default_agg.eg.go
@@ -18,7 +18,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colexecerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/colexecop"
 	"github.com/cockroachdb/cockroach/pkg/sql/colmem"
-	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
+	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 )
@@ -153,7 +153,7 @@ func (a *defaultOrderedAgg) Reset() {
 
 func newDefaultOrderedAggAlloc(
 	allocator *colmem.Allocator,
-	constructor execinfrapb.AggregateConstructor,
+	constructor execinfra.AggregateConstructor,
 	evalCtx *tree.EvalContext,
 	inputArgsConverter *colconv.VecToDatumConverter,
 	numArguments int,
@@ -183,7 +183,7 @@ type defaultOrderedAggAlloc struct {
 	aggAllocBase
 	aggFuncs []defaultOrderedAgg
 
-	constructor execinfrapb.AggregateConstructor
+	constructor execinfra.AggregateConstructor
 	evalCtx     *tree.EvalContext
 	// inputArgsConverter is a converter from coldata.Vecs to tree.Datums that
 	// is shared among all aggregate functions and is managed by the aggregator
