+++ pkg/sql/execinfrapb/processors.go
@@ -17,10 +17,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/builtins"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree/treewindow"
-	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/duration"
 	"github.com/cockroachdb/errors"
 )
@@ -34,92 +32,6 @@ func GetAggregateFuncIdx(funcName string) (int32, error) {
 		return 0, errors.Errorf("unknown aggregate %s", funcStr)
 	}
 	return funcIdx, nil
-}
-
-// AggregateConstructor is a function that creates an aggregate function.
-type AggregateConstructor func(*tree.EvalContext, tree.Datums) tree.AggregateFunc
-
-// GetAggregateInfo returns the aggregate constructor and the return type for
-// the given aggregate function when applied on the given type.
-func GetAggregateInfo(
-	fn AggregatorSpec_Func, inputTypes ...*types.T,
-) (aggregateConstructor AggregateConstructor, returnType *types.T, err error) {
-	if fn == AnyNotNull {
-		// The ANY_NOT_NULL builtin does not have a fixed return type;
-		// handle it separately.
-		if len(inputTypes) != 1 {
-			return nil, nil, errors.Errorf("any_not_null aggregate needs 1 input")
-		}
-		return builtins.NewAnyNotNullAggregate, inputTypes[0], nil
-	}
-
-	props, builtins := builtins.GetBuiltinProperties(strings.ToLower(fn.String()))
-	for _, b := range builtins {
-		typs := b.Types.Types()
-		if len(typs) != len(inputTypes) {
-			continue
-		}
-		match := true
-		for i, t := range typs {
-			if !inputTypes[i].Equivalent(t) {
-				if props.NullableArgs && inputTypes[i].IsAmbiguous() {
-					continue
-				}
-				match = false
-				break
-			}
-		}
-		if match {
-			// Found!
-			constructAgg := func(evalCtx *tree.EvalContext, arguments tree.Datums) tree.AggregateFunc {
-				return b.AggregateFunc(inputTypes, evalCtx, arguments)
-			}
-			colTyp := b.InferReturnTypeFromInputArgTypes(inputTypes)
-			return constructAgg, colTyp, nil
-		}
-	}
-	return nil, nil, errors.Errorf(
-		"no builtin aggregate for %s on %+v", fn, inputTypes,
-	)
-}
-
-// GetAggregateConstructor processes the specification of a single aggregate
-// function.
-//
-// evalCtx will not be mutated.
-func GetAggregateConstructor(
-	evalCtx *tree.EvalContext,
-	semaCtx *tree.SemaContext,
-	aggInfo *AggregatorSpec_Aggregation,
-	inputTypes []*types.T,
-) (constructor AggregateConstructor, arguments tree.Datums, outputType *types.T, err error) {
-	argTypes := make([]*types.T, len(aggInfo.ColIdx)+len(aggInfo.Arguments))
-	for j, c := range aggInfo.ColIdx {
-		if c >= uint32(len(inputTypes)) {
-			err = errors.Errorf("ColIdx out of range (%d)", aggInfo.ColIdx)
-			return
-		}
-		argTypes[j] = inputTypes[c]
-	}
-	arguments = make(tree.Datums, len(aggInfo.Arguments))
-	var d tree.Datum
-	for j, argument := range aggInfo.Arguments {
-		h := ExprHelper{}
-		// Pass nil types and row - there are no variables in these expressions.
-		if err = h.Init(argument, nil /* types */, semaCtx, evalCtx); err != nil {
-			err = errors.Wrapf(err, "%s", argument)
-			return
-		}
-		d, err = h.Eval(nil /* row */)
-		if err != nil {
-			err = errors.Wrapf(err, "%s", argument)
-			return
-		}
-		argTypes[len(aggInfo.ColIdx)+j] = d.ResolvedType()
-		arguments[j] = d
-	}
-	constructor, outputType, err = GetAggregateInfo(aggInfo.Func, argTypes...)
-	return
 }
 
 // Equals returns true if two aggregation specifiers are identical (and thus
@@ -180,60 +92,6 @@ func GetWindowFuncIdx(funcName string) (int32, error) {
 		return 0, errors.Errorf("unknown window function %s", funcStr)
 	}
 	return funcIdx, nil
-}
-
-// GetWindowFunctionInfo returns windowFunc constructor and the return type
-// when given fn is applied to given inputTypes.
-func GetWindowFunctionInfo(
-	fn WindowerSpec_Func, inputTypes ...*types.T,
-) (windowConstructor func(*tree.EvalContext) tree.WindowFunc, returnType *types.T, err error) {
-	if fn.AggregateFunc != nil && *fn.AggregateFunc == AnyNotNull {
-		// The ANY_NOT_NULL builtin does not have a fixed return type;
-		// handle it separately.
-		if len(inputTypes) != 1 {
-			return nil, nil, errors.Errorf("any_not_null aggregate needs 1 input")
-		}
-		return builtins.NewAggregateWindowFunc(builtins.NewAnyNotNullAggregate), inputTypes[0], nil
-	}
-
-	var funcStr string
-	if fn.AggregateFunc != nil {
-		funcStr = fn.AggregateFunc.String()
-	} else if fn.WindowFunc != nil {
-		funcStr = fn.WindowFunc.String()
-	} else {
-		return nil, nil, errors.Errorf(
-			"function is neither an aggregate nor a window function",
-		)
-	}
-	props, builtins := builtins.GetBuiltinProperties(strings.ToLower(funcStr))
-	for _, b := range builtins {
-		typs := b.Types.Types()
-		if len(typs) != len(inputTypes) {
-			continue
-		}
-		match := true
-		for i, t := range typs {
-			if !inputTypes[i].Equivalent(t) {
-				if props.NullableArgs && inputTypes[i].IsAmbiguous() {
-					continue
-				}
-				match = false
-				break
-			}
-		}
-		if match {
-			// Found!
-			constructAgg := func(evalCtx *tree.EvalContext) tree.WindowFunc {
-				return b.WindowFunc(inputTypes, evalCtx)
-			}
-			colTyp := b.InferReturnTypeFromInputArgTypes(inputTypes)
-			return constructAgg, colTyp, nil
-		}
-	}
-	return nil, nil, errors.Errorf(
-		"no builtin aggregate/window function for %s on %v", funcStr, inputTypes,
-	)
 }
 
 func (spec *WindowerSpec_Frame_Mode) initFromAST(w treewindow.WindowFrameMode) error {
